test and modify type

+++ src/main/java/com/asamblea/api/reactive/apirestreactivemongo/controllers/dto/MembersDto.java
@@ -1,11 +1,4 @@
 package com.asamblea.api.reactive.apirestreactivemongo.controllers.dto;
 
-public record MembersDto(Integer id,
-         String name,
-         String nameSecond,
-         String tower,
-         String apart,
-         float coefficient,
-         boolean canVote,
-         String codeBar) {
+public record MembersDto(Integer id, String name, String nameSecond, String tower, String apart, double coefficient, boolean canVote, String codeBar) {
 }
